+++ src/containerapp/azext_containerapp/_constants.py
@@ -30,7 +30,7 @@ PRIVATE_CERTIFICATE_RT = "certificates"
 DAPR_SUPPORTED_STATESTORE_DEV_SERVICE_LIST = ["postgres", "redis"]
 DAPR_SUPPORTED_PUBSUB_DEV_SERVICE_LIST = ["kafka", "redis"]
 
-DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb", "qdrant"]
+DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb", "qdrant", "milvus"]
 
 DEV_KAFKA_IMAGE = 'kafka'
 DEV_KAFKA_SERVICE_TYPE = 'kafka'
@@ -51,6 +51,10 @@ DEV_MARIADB_CONTAINER_NAME = 'mariadb'
 DEV_QDRANT_IMAGE = 'qdrant'
 DEV_QDRANT_SERVICE_TYPE = 'qdrant'
 DEV_QDRANT_CONTAINER_NAME = 'qdrant'
+
+DEV_MILVUS_IMAGE = 'milvus'
+DEV_MILVUS_SERVICE_TYPE = 'milvus'
+DEV_MILVUS_CONTAINER_NAME = 'milvus'
 
 PENDING_STATUS = "Pending"
 SUCCEEDED_STATUS = "Succeeded"